pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_hub"
version = "0.1.0"
authors = [
  {name = "Amey Desai", email = "amey.desai@nexla.com"},
  {name = "Saksham Mittal", email = "saksham.mittal@nexla.com"},
]
description = "Unified interface for major LLM providers with tracing, cost tracking, retries, structured outputs, tool calling, streaming, and more."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

dependencies = [
  "openai>=1.40",             # official OpenAI SDK (Responses API, batches, streaming)
  "anthropic>=0.34",          # official Anthropic SDK (messages, batches, streaming)
  "google-genai>=1.4",        # official Google Gen AI SDK (Gemini) — sync & async clients
  "pydantic>=2.7",
  "httpx>=0.27",
  "anyio>=4.4",
  "typing-extensions>=4.10",
  "opentelemetry-api>=1.25",  # optional tracing
  "opentelemetry-sdk>=1.25",  # optional tracing
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "mypy>=1.10",
]
mcp = [
  "mcp>=1.2",    # official Model Context Protocol Python SDK
]

[tool.ruff]
line-length = 100 